requirements.txt
flask
flask-cors
python-dotenv
gunicorn
werkzeug
